refactor(rule): split cross-validation comparators into subfolder by domain
Move comparators from a single 895-line file into cross-validation/ subfolder
with field, waste, and document comparator modules. Remove unnecessary
section-divider and restating-the-obvious comments across all files.
